requirements.txt
# Core Framework
flask>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0

# AI Integration
groq>=0.4.0
google-generativeai>=0.3.0

# Kubernetes
kubernetes>=28.1.0

# Database & Vector Search
qdrant-client>=1.7.0

# ML & Data (for anomaly detection)
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Optional: Better embeddings
# sentence-transformers>=2.2.0
